+++ egui_glium/examples/native_texture.rs
@@ -6,7 +6,7 @@ fn main() {
     let event_loop = glutin::event_loop::EventLoop::with_user_event();
     let display = create_display(&event_loop);
 
-    let mut egui_glium = egui_glium::EguiGlium::new(&display);
+    let mut egui_glium = egui_glium::EguiGlium::new(&display, &event_loop);
 
     let png_data = include_bytes!("../../examples/retained_image/src/rust-logo-256x256.png");
     let image = load_glium_image(png_data);